refactor: extract lru-cache shim
